Rewrite.
